bootstrap: Add readme, support LUKS, gen hwconfig

+++ bootstrap/install.sh
@@ -12,8 +12,18 @@ if [ ! -w "/mnt" ]; then
   exit 1
 fi
 
-cfg_root="$(dirname "$0")/.."
+cfg_root="$(realpath "$(dirname "$0")/..")"
+
+if [ ! -d "$cfg_root/hosts/$host" ]; then
+  echo "Host directory with name $host not found in $cfg_root/hosts" >&2
+  exit 1
+fi
 
 rm -rf /mnt/cfg
+echo "Copying NixOS configuration to /mnt/cfg" >&2
 cp -r "$cfg_root" /mnt/cfg
-nix-shell "/mnt/cfg/bootstrap/shell.nix" --run "nixos-install --flake /mnt/cfg#$host --show-trace"
+echo "Generating hardware configuration and installing to /mnt" >&2
+nix-shell "/mnt/cfg/bootstrap/shell.nix" --run "\
+  nixos-generate-config --root /mnt --show-hardware-config > '/mnt/cfg/hosts/$host/hardware-configuration.nix'
+  nixos-install --flake '/mnt/cfg#$host' --show-trace
+"
